+++ docs/CHANGELOG.md
@@ -1,3 +1,12 @@
+## ShowDemo 0.1.1
+
+* Show-Demo now supports -AutoPlay/-PauseBetweenStep (#39)
+* Export-Demo - Defaults to English when invariant culture (Fixes #37)
+* Improvements in how steps are determined (#35 #36)
+* Please Sponsor ShowDemo (#38)
+
+---
+
 ## ShowDemo 0.1 
 
 Initial Release of Show-Demo.
@@ -6,5 +15,3 @@ Initial Release of Show-Demo.
 * Show Demos with Show-Demo
 * Export Demos with Export-Demo.
 * ShowDemo GitHub Action
-
-